requirements.txt
opencv-python-headless==4.5.5.64
face-recognition==1.3.0
numpy==1.21.5
Pillow==9.0.1
Flask==2.0.1
Werkzeug==2.0.3
dlib==19.22.0